\name{birds}
\alias{birds}
\docType{data}
\title{
Birds in Paramo Vegetation Data
}
\description{
The data were originally derived from a study 
which investigated numbers of bird species in isolated
islands of paramo vegetation in the northern Andes.
}
\usage{data(birds)}
\format{
  A data frame with 14 observations on the following 5 variables.
  \describe{
    \item{\code{N}}{number of species}
    \item{\code{AR}}{area of island in thousands of square km}
    \item{\code{EL}}{elevation in thousands of m}
    \item{\code{Dec}}{distance from Ecuador in km}
    \item{\code{DNI}}{distance to the nearest island in km}
  }
}
\source{

   F. Vuilleumier (1970), Insular biogeography in continental regions. 
   I. The northern Andes of South America. \emph{The American Naturalist} 
   \bold{104}, 373--388

}
\keyword{datasets}
