\name{EFT}
\alias{EFT}
\docType{data}
\title{
Embedded Figures Test Data
}
\description{
Embedded figures test for 24 school children.
}
\usage{data(EFT)}
\format{
  A data frame with 24 observations on the following 3 variables.
  \describe{
    \item{\code{group}}{a factor with levels \code{row} \code{corner}}
    \item{\code{time}}{time to complete the pattern}
    \item{\code{EFT}}{Embedded Figures Test}
  }
}
\source{

  M. Aitkin, D. Anderson, B. Francis, and J. Hinde (1989),
  \emph{Statistical Modelling in GLIM}, Oxford University Press,
  New York, USA

}
\keyword{datasets}
