\name{household}
\alias{household}
\docType{data}
\title{ Household Expenditure Data }
\description{

  Survey data on household expenditure on four commodity groups.

}
\usage{data("household")}
\format{
  A data frame with 40 observations on the following 5 variables.
  \describe{
    \item{\code{housing}}{expenditure on housing, including fuel and light.}
    \item{\code{food}}{expenditure on foodstuffs, including alcohol and tobacco.}
    \item{\code{goods}}{expenditure on other goods, including clothing, footwear and durable goods.}
    \item{\code{service}}{expenditure on services, including transport and vehicles.}
    \item{\code{gender}}{a factor with levels \code{female} and \code{male}}
  }
}
\details{

  The data are part of a data set collected from a survey of household expenditure 
  and give the expenditure of 20 single men and 20 single women on four commodity groups. 
  The units of expenditure are Hong Kong dollars,

}
\source{
  FIXME
}
\examples{

  data("household", package = "HSAUR3")

}
\keyword{datasets}
