\name{Smoking_DollHill1950}
\alias{Smoking_DollHill1950}
\docType{data}
\title{
Smoking and Lung Cancer (IV)
}
\description{
Number of smokers in a case-control study.
}
\usage{data(Smoking_DollHill1950)}
\format{
  The format is:
 table [1:6, 1:2, 1:2] 2 33 250 196 136 32 27 55 293 190 ...
 - attr(*, "dimnames")=List of 3
  ..$ Smoking  : chr [1:6] "Nonsmoker" "1-" "5-" "15-" ...
  ..$ Diagnosis: chr [1:2] "Lung cancer" "Other"
  ..$ Sex      : chr [1:2] "Male" "Female"
}
\details{
This is Table V from Doll and Hill (1950).
}
\source{

  Richard Doll and A. Bradford Hill (1950), Smoking and Carcinoma of the Lung,
  \emph{British Medical Journal}, \bold{2}, 739-748

}
\references{

  Richard Doll (1998), Uncovering the effects of smoking:
  historical perspective. \emph{Statistical Methods in Medical Research},
  \bold{7}(87), 87-117

}
\keyword{datasets}
