\name{Smoking_SchairerSchoeniger1944}
\alias{Smoking_SchairerSchoeniger1944}
\docType{data}
\title{
Smoking and Lung Cancer (II)
}
\description{
Number of smokers in a case-control study.
}
\usage{data(Smoking_SchairerSchoeniger1944)}
\format{
  The format is:
 table [1:5, 1:7] 3 11 31 19 29 2 0 4 6 3 ...
 - attr(*, "dimnames")=List of 2
  ..$ Smoking  : chr [1:5] "Nonsmoker" "Moderate smoker" "Medium smoker" "Heavy smoker" ...
  ..$ Diagnosis: chr [1:7] "Lung cancer" "Lip cancer" "Throat cancer" "Stomach cancer" ...
}
\source{

  E. Schairer and E. Sch\"oninger (1944), Lungenkrebs und Tabakverbrauch,
  \emph{Zeitschrift fuer Krebsforschung}, \bold{54}(4), 261-269

}
\references{

  Richard Doll (1998), Uncovering the effects of smoking:
  historical perspective. \emph{Statistical Methods in Medical Research},
  \bold{7}(87), 87-117

}
\keyword{datasets}
