\name{Smoking_Wassink1945}
\alias{Smoking_Wassink1945}
\docType{data}
\title{
Smoking and Lung Cancer (III)
}
\description{
Number of smokers in a case-control study.
}
\usage{data(Smoking_Wassink1945)}
\format{
  The format is:
 table [1:4, 1:2] 6 18 36 74 19 36 25 20
 - attr(*, "dimnames")=List of 2
  ..$ Smoking  : chr [1:4] "Nonsmoker" "Moderate smoker" "Heavy smoker" "Very heavy smoker"
  ..$ Diagnosis: chr [1:2] "Lung cancer" "Healthy control"
}
\source{

  W. F. Wassink (1945), Ontstaansvoorwaarden voor Longkanker,
  \emph{Nederlands Tijdschrift voor Geneeskunde}, \bold{92},
  3732--3747

}
\references{

  Richard Doll (1998), Uncovering the effects of smoking:
  historical perspective. \emph{Statistical Methods in Medical Research},
  \bold{7}(87), 87-117

}
\keyword{datasets}
