\name{suicides2}
\alias{suicides2}
\docType{data}
\title{ Male Suicides Data }
\description{
  Number of suicides in different age groups and countries.
}
\usage{data("suicides2")}
\format{
  A data frame with 15 observations on the following 5 variables.
  \describe{
    \item{\code{A25.34}}{number of suicides (per 100000 males) between 25 and 34 years old.}
    \item{\code{A35.44}}{number of suicides (per 100000 males) between 35 and 44 years old.}
    \item{\code{A45.54}}{number of suicides (per 100000 males) between 45 and 54 years old.}
    \item{\code{A55.64}}{number of suicides (per 100000 males) between 55 and 64 years old.}
    \item{\code{A65.74}}{number of suicides (per 100000 males) between 65 and 74 years old.}
  }
}
\details{
  Each of the numbers gives the number of suicides per 100000 male inhabitants
  of the countries given by the row names.
}
\keyword{datasets}
