\name{USstates}
\alias{USstates}
\docType{data}
\title{ US States }
\description{
  Socio-demographic variables for ten US states.
}
\usage{data(USstates)}
\format{
  A data frame with 10 observations on the following 7 variables.
  \describe{
    \item{\code{Population}}{population size divided by 1000}
    \item{\code{Income}}{average per capita income}
    \item{\code{Illiteracy}}{illiteracy rate (per cent population)}
    \item{\code{Life.Expectancy}}{life expectancy (years)}
    \item{\code{Homicide}}{homicide rate (per 1000)}
    \item{\code{Graduates}}{percentage of high school graduates}
    \item{\code{Freezing}}{average number of days per below freezing}
  }
}
\details{

The data set contains values of seven socio-demographic 
variables for ten states in the USA.

}
\keyword{datasets}
