\name{UStemp}
\alias{UStemp}
\docType{data}
\title{ Temperatures in 22 US cities
}
\description{
Lowest temperatures in Fahrenheit in 22 US cities in four months.
}
\usage{data(UStemp)}
\format{
  A data frame with 22 observations on the following 4 variables.
  \describe{
    \item{\code{January}}{lowest temperature in Fahrenheit}
    \item{\code{April}}{lowest temperature in Fahrenheit}
    \item{\code{July}}{lowest temperature in Fahrenheit}
    \item{\code{October}}{lowest temperature in Fahrenheit}
  }
}
\keyword{datasets}
